requirements.txt
pandas>=1.1.5
sklearn>=0.0
scikit-learn>=0.24.2
chardet>=3.0.4
pytest>=6.2.5
numpy>=1.19.5
seaborn>=0.11.2
nbformat>=5.1.3
nbparameterise>=0.5
PyYAML>=6.0
click>=8.0.3
lib>=4.0.0